SCRS-5928 Linked first 3 questions in labour compliance together and added all tests

+++ app/models/api/VatComplianceLabour.scala
@@ -19,7 +19,7 @@ package models.api
 import play.api.libs.json._
 
 case class VatComplianceLabour(labour: Boolean,
-                               workers: Option[Long] = None,
+                               workers: Option[Int] = None,
                                temporaryContracts: Option[Boolean] = None,
                                skilledWorkers: Option[Boolean] = None
                               )
